new version

+++ src/modules/project/project.service.ts
@@ -21,9 +21,19 @@ export class ProjectService {
   }
   async remove(id: string): Promise<void> {
     const project = await this.projectRespository.findOneById(id);
+
     if (project?.image) {
-      const filepath = path.join(__dirname, '..', '..', '..', project.image);
-      fs.promises.unlink(filepath);
+      const isUrl =
+        project.image.startsWith('http://') ||
+        project.image.startsWith('https://');
+      if (!isUrl) {
+        const filepath = path.join(__dirname, '..', '..', '..', project.image);
+        try {
+          await fs.promises.unlink(filepath);
+        } catch {
+          console.log('image is not exits');
+        }
+      }
     }
 
     await this.projectRespository.delete(id);
